requirements.txt
asgiref==3.8.1
boto3==1.35.10
botocore==1.35.10
Django==4.2.15
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-storages==1.14.4
djangorestframework==3.15.2
jmespath==1.0.1
pillow==10.4.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
